+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Next
 
 - Add support for DC/OS 1.9 on Linux hosts.
 - ``dcos-docker doctor`` returns a status code of ``1`` if there are any errors.
+- Add a new ``dcos-docker doctor`` check for free space in the Docker root directory.
 
 2018.03.26.0
 ------------